+++ Core/Input/Mouse.h
@@ -1,23 +1,22 @@
 #pragma once
 #include "InputDevice.h"
-
 class Mouse :
     public InputDevice
 {
 private:
-    unsigned int m_X;
-    unsigned int m_Y;
+    uint64_t m_X;
+    uint64_t m_Y;
 
     bool m_IsRawMotionSupported;
 
-    int CheckForRawMotionSupport();
+    int64_t CheckForRawMotionSupport();
 public:
     Mouse() : InputDevice(), m_X{ 0 }, m_Y{ 0 }, m_IsRawMotionSupported { false } {}
 
-    unsigned int GetX() const { return m_X; }
-    unsigned int GetY() const { return m_Y; }
+    uint64_t GetX() const { return m_X; }
+    uint64_t GetY() const { return m_Y; }
 
-    int CheckForRawInputSupport();
+    int64_t CheckForRawInputSupport();
 
     //static void cursor_position_callback(GLFWwindow* window, double x_pos, double y_pos);
     //static void mouse_scroll_callback(GLFWwindow* window, double xoffset, double yoffest);
